refactor: InputAmount 객체를 로그성 대신 레파지토리 역할을 부여
