+++ docs/README.md
@@ -1,21 +1,39 @@
 ### 구현 기능 목록
 -[x] 입력금액에 맞는 동전 랜덤 생성
     - 사용자의 투입 금액으로는 동전을 생성하지 않는다.
+  
 -[x] 자판기가 보유한 동전 보여주기
+ 
 -[x] 상품등록
     - 상품명, 가격, 수량으로 상품등록
     - 상품 가격은 100원 부터
     - 10원 단위
+  
 -[x] 금액투입
+ 
 -[x] 상품구매
+ 
 -[x] 투입금액(잔여금액) 보여주기
--[] 잔돈반환 
+ 
+-[x] 잔돈을 동전으로 교환
     - 남은 금액이 상품의 최저 가격보다 적거나, 모든 상품이 소진된 경우 바로 잔돈을 돌려준다.
     - 최소 개수의 동전으로 되돌려준다.
     - 지폐를 잔돈으로 반환하는 경우는 없다.
     - 잔돈을 반환할 수 없는 경우 잔돈으로 반환할 수 있는 금액만 반환한다.
--[x] 잔돈보여주기
+  
+-[x] 잔돈출력
 
 ### 입력값 예외 케이스
--[ ] 자판기 금액 입력값은 10이상의 숫자
--[ ] 자판기 금액 입력값은 10의 배수
+-[x] 모든 입력값이 빈값일 경우 예외발생
+-[x] 자판기 금액 입력값은 숫자만 가능하다.
+-[x] 자판기 금액 입력값은 10이상의 숫자만 가능하다.
+-[x] 자판기 금액 입력값은 10의 배수만 가능하다.
+-[x] 상품 등록시 상품당 상품명, 가격, 수량 총 3가지로 이뤄져야한다.
+-[x] 상품 등록시 상품내용은 " \[ ", " \] "로 감싸져 있어야 한다.
+-[x] 상품 이름은 한 글자 이상
+-[x] 상품 가격은 숫자
+-[x] 상품 가격은 100원 이상
+-[x] 상품 가격은 10원 단위
+-[x] 상품 수량은 숫자
+-[x] 상품 수량은 1개 이상
+-[x] 상품 구매시 등록된 상품인지 확인
