+++ src/main/java/view/View.java
@@ -1,14 +1,13 @@
-package vendingmachine;
+package view;
 
 import java.util.Map;
 
 import camp.nextstep.edu.missionutils.Console;
 import vendingmachine.domain.Coin;
-import vendingmachine.repository.InputAmount;
 import vendingmachine.utils.ValidationUtil;
 
 public class View {
-    public static void printVendingMachineCoin(Map<Coin, Integer> coins) {
+    public static void printInitialCoins(Map<Coin, Integer> coins) {
         System.out.println();
         System.out.println("자판기가 보유한 동전");
         System.out.println("500원 - " + coins.getOrDefault(Coin.COIN_500, 0) + "개");
@@ -17,7 +16,7 @@ public class View {
         System.out.println("10원 - " + coins.getOrDefault(Coin.COIN_10, 0) + "개");
     }
 
-    public static void printCoins(InputAmount inputAmount, Map<Coin, Integer> coins) {
+    public static void printCoins(int inputAmount, Map<Coin, Integer> coins) {
         System.out.println();
         System.out.println("투입 금액: " + inputAmount + "원");
         System.out.println("잔돈");
@@ -73,7 +72,7 @@ public class View {
         return inputAmount();
     }
 
-    public static String buyProduct(InputAmount userBalance) {
+    public static String buyProduct(int userBalance) {
         System.out.println();
         System.out.println("투입 금액: " + userBalance + "원");
         System.out.println("구매할 상품명을 입력해 주세요.");
